Fix multiplication-dereference ambiguity
Through an external function in the grammar I can lookahead in the
`mul_expr` rule to fail it when it should be a `deref_expr` instead.
However, the current function does a basic check for some common cases
to match part of the current cppfront functionality.

I doubt cppfront is working exactly as intended in all of these cases
so fixing it in a more complete way will be deferred until I have made
sure how cppfront is meant to work in these cases.

Closes #14.