Fix grammar/lexer to handle comments correctly
I have been using comments as if they were just another AST/PSI token,
and it meant I would have to add line and block comments to all places
in all rules to handle it, which is absurd.

It could be fixed as the
`PsiBuilder` apparently ignores them (it seems it's based on
`ParserDefinition.getCommentTokens()`) by default. However, my
`getCommentTokens()` was not set up correctly, and it also requires
the lexeme definitions of the elements used here to be specified in the
`.bnf file`, as well.